Bug #1668107 Fix PC-lint bug
[root cause  ] Wrong data type
[changes     ] libcamera
[side effects] no
[self test   ]: yes
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: yes
[board]: ums9230
[test case]: yes
[reviewers   ] victor.xu
[change_type ] feature_bugfix
[tag_product ] common

+++ hal3_2v6/SprdCamera3Setting.cpp
@@ -4720,7 +4720,7 @@ int SprdCamera3Setting::updateAppMode(
         if (s_setting[mCameraId].flash_InfoInfo.available == 0 &&
             !strcmp(FRONT_CAMERA_FLASH_TYPE, "lcd") &&
             cameraInfo.facing == CAMERA_FACING_FRONT) {
-            int32_t sprd_flash_lcd_mode =
+            uint8_t sprd_flash_lcd_mode =
                 frame_settings.find(ANDROID_SPRD_FLASH_LCD_MODE).data.u8[0];
             GET_VALUE_IF_DIF(
                 s_setting[mCameraId].sprddefInfo.sprd_flash_lcd_mode,
@@ -5117,7 +5117,7 @@ int SprdCamera3Setting::updateWorkParameters(
         if (s_setting[mCameraId].flash_InfoInfo.available == 0 &&
             !strcmp(FRONT_CAMERA_FLASH_TYPE, "lcd") &&
             cameraInfo.facing == CAMERA_FACING_FRONT) {
-            int32_t sprd_flash_lcd_mode =
+            uint8_t sprd_flash_lcd_mode =
                 frame_settings.find(ANDROID_SPRD_FLASH_LCD_MODE).data.u8[0];
             GET_VALUE_IF_DIF(
                 s_setting[mCameraId].sprddefInfo.sprd_flash_lcd_mode,
